==> 06.TextAnalysis/01.BagofWords.ipynb <==
## Bag Of Words

### 1. Count Vectorizer

from sklearn.feature_extraction.text import CountVectorizer
corpus = ['you know I want your love. because I love you.']

count_vect = CountVectorizer()
output = count_vect.fit_transform(corpus)
output.toarray()

count_vect.vocabulary_

- 불용어를 제거한 BoW

# 자체 제거
text = ["Family is not an important thing. It's everything."]
cvect = CountVectorizer(stop_words=['the','a','an','is'])
print(cvect.fit_transform(text).toarray())
print(cvect.vocabulary_)

# Scikit-Learn에서 제공하는 불용어 사용
text = ["Family is not an important family thing. It's everything."]
cvect = CountVectorizer(stop_words='english')
print(cvect.fit_transform(text).toarray())
print(cvect.vocabulary_)

# NLTK에서 제공하는 불용어 사용
from nltk.corpus import stopwords
sw = stopwords.words('english')
len(sw)


text = ["Family is not an important thing. It's everything."]
cvect = CountVectorizer(stop_words=sw)
print(cvect.fit_transform(text).toarray())
print(cvect.vocabulary_)

def get_word(index, voca):
    for key,value in voca.items():
        

- N-gram

text = ["Machine learning is fun and not boring"]
cvect = CountVectorizer()
print(cvect.fit_transform(text).toarray())
print(cvect.vocabulary_)

cvect = CountVectorizer(ngram_range=(1,2))
print(cvect.fit_transform(text).toarray())
print(cvect.vocabulary_)

cvect = CountVectorizer(ngram_range=(1,3),stop_words='english')
print(cvect.fit_transform(text).toarray())
print(cvect.vocabulary_)

- Hyper Parameter

cvect.get_params()

### 2. TF-IDF Vectorizer

from sklearn.feature_extraction.text import TfidfVectorizer
corpus = [
    "you know I want your love",
    "I like you",
    "what should I do"
]

cvect = CountVectorizer()
print(cvect.fit_transform(corpus).toarray())
print(cvect.vocabulary_)

tvect = TfidfVectorizer()
print(tvect.fit_transform(corpus).toarray())
print(tvect.vocabulary_)

tvect = TfidfVectorizer(ngram_range=(1,2), stop_words='english')
print(tvect.fit_transform(corpus).toarray())
print(tvect.vocabulary_)

tvect.get_params()
