==> 06.TextAnalysis/02.NewsGroup.ipynb <==
## 20 뉴스 그룹 분류

import numpy as np
import pandas as pd

from sklearn.datasets import fetch_20newsgroups
news_data = fetch_20newsgroups(subset='all',random_state=156)

- 데이터 탐색

news_data.keys()

pd.Series(news_data.target).value_counts().sort_index()

print(news_data.data[0])

- 훈련/테스트용 데이터 추출

train_news = fetch_20newsgroups(subset='train', random_state=150, remove=('headers','footers','quotes'))
X_train = train_news.data
y_train = train_news.target

print(train_news.data[0])

train_news.target[0], train_news.target_names[train_news.target[0]]

test_news = fetch_20newsgroups(subset='test', random_state=156, remove=('headers','footers','quotes'))
X_test = test_news.data
y_test = test_news.target

len(X_train), len(X_test)

### 피쳐 벡터화 변환과 머신러닝 모델 학습/예측/평가
- Case 1. CountVectorizer + LogisticRegression

from sklearn.feature_extraction.text import CountVectorizer
cvect = CountVectorizer()
cvect.fit(X_train)
X_train_cv = cvect.transform(X_train)
X_test_cv = cvect.transform(X_test)

X_train_cv.shape, X_test_cv.shape

from sklearn.linear_model import LogisticRegression
lr = LogisticRegression(max_iter=300)
lr.fit(X_train_cv, y_train)

lr.get_params()

from sklearn.metrics import accuracy_score
pred=lr.predict(X_test_cv)
accuracy_score(y_test,pred)

y_test[:5], pred[:5]

- Case 2. TfidVectorizer + LogisticRegression

from sklearn.feature_extraction.text import TfidfVectorizer
tvect = TfidfVectorizer()
tvect.fit(X_train)
X_train_tf = tvect.transform(X_train)
X_test_tf = tvect.transform(X_test)

X_train_tf.shape, X_test_tf.shape

lr = LogisticRegression(max_iter=300)
lr.fit(X_train_tf, y_train)

pred = lr.predict(X_test_tf)
accuracy_score(y_test,pred)

y_test[:5], pred[:5]

- Case 3. stop_words filtering, ngram_range=(1,2), max_df=300

tvect2 = TfidfVectorizer(ngram_range=(1,2), max_df=300, stop_words='english')
tvect2.fit(X_train)
X_train_tf2 = tvect2.transform(X_train)
X_test_tf2 = tvect2.transform(X_test)

X_train_tf2.shape, X_test_tf2.shape

lr = LogisticRegression(max_iter=300)
lr.fit(X_train_tf2, y_train)
pred2 = lr.predict(X_test_tf2)
accuracy_score(y_test, pred2)

- Case 4. Case 3에서 LogisticRegression C값을 10으로 

lr = LogisticRegression(max_iter=300, C=10)
lr.fit(X_train_tf2, y_train)
pred2 = lr.predict(X_test_tf2)
accuracy_score(y_test, pred2)

### Pipeline과 GridSearchCV를 통한 하이퍼 파라미터 튜닝

from sklearn.pipeline import Pipeline

pipeline = Pipeline([
    ('tvect', TfidfVectorizer(stop_words='english')),
    ('lr', LogisticRegression())
])

params = {
    'tvect__ngram_range':[(1,1),(1,2)],
    'tvect__max_df':[300,700],
    'lr__C':[1,10]
}

from sklearn.model_selection import GridSearchCV

grid_pipe = GridSearchCV(
    pipeline, param_grid=params, cv=3, scoring='accuracy', verbose=1, n_jobs=-1
    )

%time grid_pipe.fit(X_train,y_train)

print(grid_pipe.best_score_)
print(grid_pipe.best_params_)
pred = grid_pipe.best_estimator_(X_test)
accuracy_score(y_test, pred)
