==> 06.TextAnalysis/03.IMDB.ipynb <==
## IMDB 영화평 감성분석

import pandas as pd
import numpy as np

df = pd.read_csv('labeledTrainData.tsv', sep='\t', quoting=3)
df.head()

df.info()

df.review[0][:1000]

### 전처리

# <br />태그는 공백으로 변환
df['review'] = df.review.str.replace('<br />',' ')

# 구둣점, 숫자 제거 --> 영어 이외의 문자는 공백으로 변환
import re
df['review'] = df.review.apply(lambda x: re.sub('[^A-Za-z]', ' ', x))

df.review[0][:1000]

### train/test data set 분리 및 텍스트 변환

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(
    df.review, df.sentiment, stratify=df.sentiment, random_state=2021
)

X_train.shape, X_test.shape, y_train.shape, y_test.shape

- CountVectorizer

from sklearn.feature_extraction.text import CountVectorizer
cvect = CountVectorizer(stop_words='english', ngram_range=(1,2))
cvect.fit(X_train)
X_train_cv = cvect.transform(X_train)
X_test_cv = cvect.transform(X_test)

X_train_cv.shape, X_test_cv.shape

from sklearn.neighbors import KNeighborsClassifier
knn = KNeighborsClassifier()
%time knn.fit(X_train_cv, y_train)

from sklearn.metrics import accuracy_score
pred = knn.predict(X_test_cv)
accuracy_score(y_test, pred)

from sklearn.linear_model import LogisticRegression
lr = LogisticRegression(max_iter = 300)
%time lr.fit(X_train_cv, y_train)

pred = lr.predict(X_test_cv)
accuracy_score(y_test, pred)

- TfidVectorizer

from sklearn.feature_extraction.text import TfidfVectorizer
tvect = TfidfVectorizer(stop_words='english', ngram_range=(1,2))
tvect.fit(X_train)
X_train_tf = tvect.transform(X_train)
X_test_tf = tvect.transform(X_test)

lr = LogisticRegression(max_iter=300)
%time lr.fit(X_train_tf, y_train)
pred_tf = lr.predict(X_test_tf)
accuracy_score(y_test, pred_tf)

### 모델 저장하고 불러오기

import joblib

joblib.dump(tvect, 'model/imdb_tvect.pkl')
joblib.dump(lr,'model/imdb_lr.pkl')

# 변수 지우기
del tvect
del lr

new_tvect = joblib.load('model/imdb_tvect.pkl')
new_lr = joblib.load('model/imdb_lr.pkl')

new_X_test_tf = new_tvect.transform(X_test)
new_pred = new_lr.predict(new_X_test_tf)
accuracy_score(y_test, new_pred)

### pipeline과 GridSearchCV 를 통한 하이퍼 파라미터 튜닝

from sklearn.pipeline import Pipeline
pipeline = Pipeline([
    ('tvect', TfidfVectorizer(stop_words='english',ngram_range=(1,2))),
    ('lr',LogisticRegression())
])
params = {
    'tvect__max_df': [100,500],
    'lr__C':[1,10]
}

from sklearn.model_selection import GridSearchCV
pipe_grid = GridSearchCV(
    pipeline, param_grid=params, cv=3, verbose=1, n_jobs=-1, scoring='accuracy'
)
%time pipe_grid.fit(X_train, y_train)

print(pipe_grid.best_score_)
print(pipe_grid.best_params_)

pred = pipe_grid.best_estimator_.predict(X_test)
accuracy_score(y_test, pred)

joblib.dump(pipe_grid, 'model/imdb_pipe.pkl')
