==> 06.TextAnalysis/04.naver_movie.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

# '네이버 영화 리뷰 데이터' 로 검색
train_df = pd.read_csv("https://raw.githubusercontent.com/e9t/nsmc/master/ratings_train.txt", sep='\t')
test_df = pd.read_csv("https://raw.githubusercontent.com/e9t/nsmc/master/ratings_test.txt", sep='\t')

train_df.head(3)

train_df.shape, test_df.shape

### 1. 데이터 전처리

- train data set

# 중복 여부 확인
train_df.document.nunique()

# 중복 데이터 제거
train_df.drop_duplicates(subset=['document'], inplace=True)
train_df.shape

# 결측치 있는지 확인
train_df.isnull().sum()

# 결측치 제거
train_df = train_df.dropna(how='any')
train_df.shape

train_df.label.value_counts()  # 확인

- test data set

# 중복 여부 확인
test_df.document.nunique()

# 중복 데이터 제거
test_df.drop_duplicates(subset=['document'], inplace=True)
test_df.shape

# 결측치 있는지 확인
test_df.isnull().sum()

# 결측치 제거
test_df = test_df.dropna(how='any')
test_df.shape   # 확인

### 2. 텍스트 전처리

- train data set

import warnings
warnings.filterwarnings(action='ignore')

# 한글과 공백 이외는 제거
train_df['document'] = train_df.document.str.replace('[^ㄱ-ㅎㅏ-ㅣ가-힣 ]','')
train_df.head(3)

# '' 데이터는 Nan으로 변환한 후 제거
train_df['document'].replace('', np.nan, inplace=True)
train_df.document.isnull().sum()

# 결측치 제거
train_df = train_df.dropna(how='any')
train_df.shape

- test data set

test_df['document'] = test_df.document.str.replace('[^ㄱ-ㅎㅏ-ㅣ가-힣 ]','')

test_df['document'].replace('', np.nan, inplace=True)
test_df.document.isnull().sum()

test_df = test_df.dropna(how='any')
test_df.shape

- 새로운 파일로 저장

train_df.to_csv('naver_movie_train.tsv', sep = '\t', index = False)
test_df.to_csv('naver_movie_test_df.tsv', sep = '\t', index = False)

### 3. 한글 처리

from konlpy.tag import Okt
okt = Okt()

stopwords = ['은','는','이','가','의','들','좀','잘','걍','과','도','을','를','으로','자','에','와','한','하다']

text = '교도소 이야기구먼 솔직히 지루하고 재미는 없음평점 조정해야할듯'
okt.morphs(text)

okt.morphs(text, stem=True)

from tqdm import tqdm_notebook
tqdm_notebook

# X_train
from tqdm import tqdm_notebook
X_train = []

for sentence in tqdm_notebook(train_df.document):
    morphs = okt.morphs(sentence, stem=True)
    temp_X = ' '.join([word for word in morphs if not word in stopwords])
    X_train.append(temp_X)

# X_test
from tqdm import tqdm_notebook
X_test = []

for sentence in tqdm_notebook(test_df.document):
    morphs = okt.morphs(sentence, stem=True)
    temp_X = ' '.join([word for word in morphs if not word in stopwords])
    X_test.append(temp_X)

y_train = train_df.label.values
y_test = test_df.label.values

X_train[:5]

### 4. Feature 변환, 모델 학습/예측/평가

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score

cvect = CountVectorizer()
cvect.fit(X_train)
X_train_cv = cvect.transform(X_train)
X_test_cv = cvect.transform(X_test)

lr = LogisticRegression()
lr.fit(X_train_cv, y_train)

pred = lr.predict(X_test_cv)
accuracy_score(y_test, pred)

### 5. 실제 테스트

review1 = ' 운 좋게 시사회로 봤는데 진짜 재밌었음.. 올해 본 영화중 최고였음 ㅠ 문화의 날로 할인받아서 엄마랑 볼려고 또 예매함ㅎ'
review2 = '그냥 미쳤다. 뭐 하나 빠지는게 없다. 101마리 달마시안 팬들은 영화가 끝난것같아도 나가지말고 일단 앉아있으세요 ㅋㅋㅋㅋ?'

review3 = '세상에.. 정말 오랜만에 보는 영화라 기대했는데 기대 이하 정도가 아니라 이 영화를 보느라 할애한 제 귀중한 시간이 아까울 정도였습니다. 유치한 신파 요소는 그렇다 쳐도 영국 배경인데 그 흔한 유머 단 한마디도 없이...'

review4='세계에서 제일 매력 있는 빌런 영화'

type(review1)

# 이렇게하면 안됨..이유가 뭘까
review1 = review1.replace('[^ㄱ-ㅎㅏ-ㅣ가-힣 ]','')
review1

# 다른 방법으로 해결
import re
review1 = re.sub('[^ㄱ-ㅎㅏ-ㅣ가-힣 ]', '', review1)
review1

morphs = okt.morphs(review1)
review = ' '.join([word for word in morphs if not word in stopwords])
review

review_cv = cvect.transform([review])  # 리스트로 만들어서 넣기
pred = lr.predict(review_cv)
print(pred)
print(pred[0])  # -- 결과: 1을 0으로 예측

# review2 에 대해 동일한 과정을 진행 -- 결과: 1을 0으로 예측

review2 = re.sub('[^ㄱ-ㅎㅏ-ㅣ가-힣 ]', '', review2)
morphs = okt.morphs(review2)
review = ' '.join([word for word in morphs if not word in stopwords])
review_cv = cvect.transform([review]) 
pred = lr.predict(review_cv)
print(pred)
print(pred[0])

# review3 에 대해 동일한 과정을 진행 -- 결과: 0을 0으로 예측

review3 = re.sub('[^ㄱ-ㅎㅏ-ㅣ가-힣 ]', '', review3)
morphs = okt.morphs(review3)
review = ' '.join([word for word in morphs if not word in stopwords])
review_cv = cvect.transform([review]) 
pred = lr.predict(review_cv)
print(pred)
print(pred[0])

# review4 에 대해 동일한 과정을 진행 -- 결과: 1을 1로 예측

review4 = re.sub('[^ㄱ-ㅎㅏ-ㅣ가-힣 ]', '', review4)
morphs = okt.morphs(review4)
review = ' '.join([word for word in morphs if not word in stopwords])
review_cv = cvect.transform([review]) 
pred = lr.predict(review_cv)
print(pred)
print(pred[0])

- 여러개의 리뷰가 리스트로 주어졌을 때

reviews = ['아름다운 음악과 아름다운 풍광~ 그렇지 못한 현실이 찡하네요~',
'메시지와 작위성의 불협화음!!!']

review_list = []
for review in reviews:
    review = re.sub('[^ㄱ-ㅎㅏ-ㅣ가-힣 ]', '', review)
    morphs = okt.morphs(review, stem=True)
    temp_X = ' '.join([word for word in morphs if not word in stopwords])
    review_list.append(temp_X)

review_list

review_cv = cvect.transform(review_list)
pred = lr.predict(review_cv)
pred

### 6. GridSearchCV로 최적 파라미터 찾기

from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV

pipeline = Pipeline([
    ('cvect', CountVectorizer()),
    ('lr', LogisticRegression())
])

params = { 'cvect__ngram_range':[(1,1),(1,2)],
            'cvect__max_df':[0.9, 0.99],
            'cvect__min_df':[1, 3],
            'lr__C':[1,5]
}

grid_pipe = GridSearchCV(
    pipeline, param_grid=params, cv=3, n_jobs=-1, scoring='accuracy'
)
%time grid_pipe.fit(X_train, y_train)
print(grid_pipe.best_params_)
print(grid_pipe.best_score_)

pred = grid_pipe.best_estimator_.predict(X_test)
accuracy_score(y_test, pred)
