==> 06.TextAnalysis/05.naver_movie2.ipynb <==
## 네이버 영화평 감성분석 - TFidfVectorizer

import pandas as pd
import numpy as np

train_df = pd.read_csv('naver_movie_train.tsv', sep='\t')
test_df = pd.read_csv('naver_movie_test_df.tsv', sep='\t')
train_df.head(3)

- Tokenizer 함수 정의

from konlpy.tag import Okt
okt = Okt()

stopwords = ['은','는','이','가','의','들','좀','잘','걍','과','도','을','를','으로','자','에','와','한','하다']

def okt_tokenizer(text):
    tokens = okt.morphs(text, stem=True)
    tokens = [word for word in tokens if not word in stopwords]
    return tokens

review1 = ' 운 좋게 시사회로 봤는데 진짜 재밌었음.. 올해 본 영화중 최고였음 ㅠ 문화의 날로 할인받아서 엄마랑 볼려고 또 예매함ㅎ'

okt_tokenizer(review1)

### TfidfVectorizer 변환

import warnings
warnings.filterwarnings(action='ignore')

from sklearn.feature_extraction.text import TfidfVectorizer

tvect = TfidfVectorizer(
    tokenizer=okt_tokenizer, ngram_range=(1,2), max_df = 0.9)

%time tvect.fit(train_df.document)

%time X_train_tv = tvect.transform(train_df.document)

X_test_tv = tvect.transform(test_df.document)

y_train = train_df.label.values
y_test = test_df.label.values

### Naive Bayes 분류기로 학습/예측/평가

from sklearn.naive_bayes import MultinomialNB
nb = MultinomialNB()
nb.get_params()

nb.fit(X_train_tv, y_train)

from sklearn.metrics import accuracy_score
pred = nb.predict(X_test_tv)
accuracy_score(y_test, pred)

### 실제 테스트

reviews = ['아름다운 음악과 아름다운 풍광~ 그렇지 못한 현실이 찡하네요~',
'메시지와 작위성의 불협화음!!!']

reviews_tv = tvect.transform(reviews)
pred = nb.predict(reviews_tv)
pred
